+++ OpenGlass/GlassSharedStructures.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+namespace OpenGlass
+{
+	enum class Type : unsigned char
+	{
+		Blur = 0,
+		Aero = 1
+	};
+}
